Add anonymous user, change Report model to use grid instead of location
